serilog loglama işlemi gerçekleştirildi. Seq ile görsel de ulaşılabilir

+++ VetProManager.DAL/Contracts/BaseContracts/IRepository.cs
@@ -2,7 +2,7 @@
 
 namespace VetProManager.DAL.Contracts.BaseContracts {
     public interface IRepository<T> where T : class {
-        Task<T?> GetByIdAsync(int Id);
+        Task<T?> GetByIdAsync(long Id);
         Task<IEnumerable<T>> GetAllAsync();
         IQueryable<T> GetAllAsQueryable();
         Task<IEnumerable<T>> Where(Expression<Func<T, bool>> predicate);
