+++ VetProManager.Service/BaseService/Contract/IService.cs
@@ -2,7 +2,7 @@
 
 namespace VetProManager.Service.BaseService.Contract {
     public interface IService<T> where T: class {
-        Task<T?> GetByIdAsync(int Id);
+        Task<T?> GetByIdAsync(long Id);
         Task<IEnumerable<T>> GetAllAsync();
         IQueryable<T> GetAllAsQueryable();
         Task<IEnumerable<T>> Where(Expression<Func<T, bool>> predicate);
